@@ -12,23 +12,23 @@
   },
   "devDependencies": {
     "@rollup/plugin-commonjs": "^20.0.0",
-    "@rollup/plugin-node-resolve": "^13.0.4",
-    "@rollup/plugin-typescript": "^8.0.0",
-    "@tsconfig/svelte": "^2.0.0",
+    "@rollup/plugin-node-resolve": "^13.0.5",
+    "@rollup/plugin-typescript": "^8.2.5",
+    "@tsconfig/svelte": "^2.0.1",
     "parse5": "^6.0.1",
-    "rollup": "^2.56.3",
+    "rollup": "^2.57.0",
     "rollup-plugin-css-only": "^3.1.0",
     "rollup-plugin-json": "^4.0.0",
     "rollup-plugin-livereload": "^2.0.5",
     "rollup-plugin-svelte": "^7.1.0",
     "rollup-plugin-terser": "^7.0.2",
-    "svelte": "^3.42.6",
-    "svelte-check": "^2.0.0",
-    "svelte-preprocess": "^4.0.0",
-    "svelte-translate": "^0.1.5",
+    "svelte": "^3.43.0",
+    "svelte-check": "^2.2.6",
+    "svelte-preprocess": "^4.9.5",
+    "svelte-translate": "^0.1.8",
     "svelte-translate-tools": "./../",
-    "tslib": "^2.0.0",
-    "typescript": "^4.0.0"
+    "tslib": "^2.3.1",
+    "typescript": "^4.4.3"
   },
   "dependencies": {
     "sirv-cli": "^1.0.14"
